% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main function.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the total number of trials}

\item{k}{the total number of successes}
}
\value{
the combinations "n choose k"
}
\description{
Caculate the combinations "n choose k"
}
\examples{
bin_choose(n = 5, k = 2)
}
