% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main function.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{the total number of trials}

\item{prob}{the probability of success every time}
}
\value{
a numeric object
}
\description{
caculate the mean of a binominal distribution
}
\examples{
bin_mean(trials = 5, prob = 0.5)
}
