% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main function.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes}

\item{trials}{the total number of trials}

\item{prob}{the probability of success every time}
}
\value{
the result of a binominal distribution.
}
\description{
To caculate the result of a binominal distribution
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)

bin_probability(success = 0:2, trials = 5, prob = 0.5)

}
