% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main function.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{the total number of trials}

\item{prob}{the probability of success every time}
}
\value{
a numeric object
}
\description{
caculate the variance of a binominal distribution
}
\examples{
bin_variance(trials = 5, prob = 0.5)
}
